requirements_weviate.txt
chainlit 
farm-haystack[weaviate,inference]
requests 
transformers 

pdfplumber  # for PDF handling

weaviate-haystack